mqtt-client/requirements.txt
# Core MQTT and WebSocket dependencies
paho-mqtt>=1.6.1
websockets>=10.3

# System monitoring and health checks
psutil>=5.9.0

# Development and testing utilities
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Enhanced logging and monitoring (optional)
structlog>=22.0.0

# Docker and container support
requests>=2.28.0

# Data validation (optional but recommended)
pydantic>=1.10.0

# Environment configuration
python-dotenv>=0.19.0